organize processing fn
